+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * add support for `django.core.cache.backends.memcached.PyMemcacheCache` as `pymemcached:` protocol
 * rename `memcached+pylibmccache` protocol to `pylibmccache`
+* fix parsing of `pylibmccache` service
 
 ## 1.3.0